+++ packages/interface/src/components/preferences-modal/index.js
@@ -7,7 +7,7 @@ import { __ } from '@wordpress/i18n';
 export default function PreferencesModal( { closeModal, children } ) {
 	return (
 		<Modal
-			className="interface-preferences-modal"
+			className="interface-preferences-modal interface-complementary-area"
 			title={ __( 'Preferences' ) }
 			onRequestClose={ closeModal }
 		>