+++ lib/load.php
@@ -41,12 +41,10 @@ if ( class_exists( 'WP_REST_Controller' ) ) {
 	require_once __DIR__ . '/compat/wordpress-6.2/class-gutenberg-rest-pattern-directory-controller-6-2.php';
 	require_once __DIR__ . '/compat/wordpress-6.2/rest-api.php';
 	require_once __DIR__ . '/compat/wordpress-6.2/block-patterns.php';
-	require_once __DIR__ . '/compat/wordpress-6.2/class-gutenberg-rest-global-styles-controller-6-2.php';
 
 	// WordPress 6.3 compat.
 	require_once __DIR__ . '/compat/wordpress-6.3/class-gutenberg-rest-block-patterns-controller-6-3.php';
 	require_once __DIR__ . '/compat/wordpress-6.3/class-gutenberg-rest-templates-controller-6-3.php';
-	require_once __DIR__ . '/compat/wordpress-6.3/class-gutenberg-rest-global-styles-controller-6-3.php';
 	require_once __DIR__ . '/compat/wordpress-6.3/class-gutenberg-rest-global-styles-revisions-controller-6-3.php';
 	require_once __DIR__ . '/compat/wordpress-6.3/class-gutenberg-classic-to-block-menu-converter.php';
 	require_once __DIR__ . '/compat/wordpress-6.3/class-gutenberg-navigation-fallback.php';
@@ -60,9 +58,12 @@ if ( class_exists( 'WP_REST_Controller' ) ) {
 	require_once __DIR__ . '/compat/wordpress-6.3/footnotes.php';
 
 	// WordPress 6.4 compat.
-	require_once __DIR__ . '/compat/wordpress-6.4/class-gutenberg-rest-global-styles-controller-6-4.php';
 	require_once __DIR__ . '/compat/wordpress-6.4/class-gutenberg-rest-global-styles-revisions-controller-6-4.php';
 	require_once __DIR__ . '/compat/wordpress-6.4/rest-api.php';
+
+	// Plugin specific code.
+	require_once __DIR__ . '/class-wp-rest-global-styles-controller-gutenberg.php';
+	require_once __DIR__ . '/rest-api.php';
 
 	// Experimental.
 	if ( ! class_exists( 'WP_Rest_Customizer_Nonces' ) ) {
@@ -120,6 +121,7 @@ require_once __DIR__ . '/compat/wordpress-6.3/kses.php';
 
 // WordPress 6.4 compat.
 require __DIR__ . '/compat/wordpress-6.4/blocks.php';
+require __DIR__ . '/compat/wordpress-6.4/block-patterns.php';
 
 // Experimental features.
 require __DIR__ . '/experimental/block-editor-settings-mobile.php';
@@ -152,37 +154,40 @@ require __DIR__ . '/experimental/interactivity-api/directives/wp-text.php';
 remove_action( 'plugins_loaded', '_wp_theme_json_webfonts_handler' ); // Turns off WordPress 6.0's stopgap handler.
 
 /*
- * If the Fonts Library is available, load the Font Face files, else load the Fonts API.
- * This strategy is temporary until the Fonts Library is merged. It's used here to allow
- * the Font Face (redesigned Fonts API) to be merged before the Fonts Library while
+ * If the Font Library is available, load the Font Face files, else load the Fonts API.
+ * This strategy is temporary until the Font Library is merged. It's used here to allow
+ * the Font Face (redesigned Fonts API) to be merged before the Font Library while
  * keeping Fonts API available for sites that are using it.
  */
-if ( defined( 'FONTS_LIBRARY_ENABLE' ) && FONTS_LIBRARY_ENABLE ) {
-	// Loads the Fonts Library.
-	require __DIR__ . '/experimental/fonts-library/class-wp-fonts-library.php';
-	require __DIR__ . '/experimental/fonts-library/class-wp-font-family-utils.php';
-	require __DIR__ . '/experimental/fonts-library/class-wp-font-family.php';
-	require __DIR__ . '/experimental/fonts-library/class-wp-rest-fonts-library-controller.php';
-	require __DIR__ . '/experimental/fonts-library/fonts-library.php';
+if (
+	( defined( 'FONT_LIBRARY_ENABLE' ) && FONT_LIBRARY_ENABLE ) ||
+	( defined( 'FONTS_LIBRARY_ENABLE' ) && FONTS_LIBRARY_ENABLE )
+) {
+	// Loads the Font Library.
+	require __DIR__ . '/experimental/fonts/font-library/class-wp-font-library.php';
+	require __DIR__ . '/experimental/fonts/font-library/class-wp-font-family-utils.php';
+	require __DIR__ . '/experimental/fonts/font-library/class-wp-font-family.php';
+	require __DIR__ . '/experimental/fonts/font-library/class-wp-rest-font-library-controller.php';
+	require __DIR__ . '/experimental/fonts/font-library/font-library.php';
 
 	if ( ! class_exists( 'WP_Font_Face' ) ) {
-		require __DIR__ . '/experimental/fonts/class-wp-font-face.php';
-		require __DIR__ . '/experimental/fonts/class-wp-font-face-resolver.php';
-		require __DIR__ . '/experimental/fonts/fonts.php';
+		require __DIR__ . '/experimental/fonts/font-face/class-wp-font-face.php';
+		require __DIR__ . '/experimental/fonts/font-face/class-wp-font-face-resolver.php';
+		require __DIR__ . '/experimental/fonts/font-face/fonts.php';
 
 		// Load the BC Layer. Do no backport to WP Core.
-		require __DIR__ . '/experimental/fonts/bc-layer/class-wp-fonts-provider.php';
-		require __DIR__ . '/experimental/fonts/bc-layer/class-wp-fonts-utils.php';
-		require __DIR__ . '/experimental/fonts/bc-layer/class-wp-fonts.php';
-		require __DIR__ . '/experimental/fonts/bc-layer/class-wp-fonts-provider-local.php';
-		require __DIR__ . '/experimental/fonts/bc-layer/class-wp-fonts-resolver.php';
-		require __DIR__ . '/experimental/fonts/bc-layer/class-gutenberg-fonts-api-bc-layer.php';
-		require __DIR__ . '/experimental/fonts/bc-layer/webfonts-deprecations.php';
-		require __DIR__ . '/experimental/fonts/bc-layer/class-wp-webfonts-utils.php';
-		require __DIR__ . '/experimental/fonts/bc-layer/class-wp-webfonts-provider.php';
-		require __DIR__ . '/experimental/fonts/bc-layer/class-wp-webfonts-provider-local.php';
-		require __DIR__ . '/experimental/fonts/bc-layer/class-wp-webfonts.php';
-		require __DIR__ . '/experimental/fonts/bc-layer/class-wp-web-fonts.php';
+		require __DIR__ . '/experimental/fonts/font-face/bc-layer/class-wp-fonts-provider.php';
+		require __DIR__ . '/experimental/fonts/font-face/bc-layer/class-wp-fonts-utils.php';
+		require __DIR__ . '/experimental/fonts/font-face/bc-layer/class-wp-fonts.php';
+		require __DIR__ . '/experimental/fonts/font-face/bc-layer/class-wp-fonts-provider-local.php';
+		require __DIR__ . '/experimental/fonts/font-face/bc-layer/class-wp-fonts-resolver.php';
+		require __DIR__ . '/experimental/fonts/font-face/bc-layer/class-gutenberg-fonts-api-bc-layer.php';
+		require __DIR__ . '/experimental/fonts/font-face/bc-layer/webfonts-deprecations.php';
+		require __DIR__ . '/experimental/fonts/font-face/bc-layer/class-wp-webfonts-utils.php';
+		require __DIR__ . '/experimental/fonts/font-face/bc-layer/class-wp-webfonts-provider.php';
+		require __DIR__ . '/experimental/fonts/font-face/bc-layer/class-wp-webfonts-provider-local.php';
+		require __DIR__ . '/experimental/fonts/font-face/bc-layer/class-wp-webfonts.php';
+		require __DIR__ . '/experimental/fonts/font-face/bc-layer/class-wp-web-fonts.php';
 	}
 } elseif ( ! class_exists( 'WP_Fonts' ) ) {
 	// Fonts API files.