+++ packages/components/src/navigator/stories/index.story.tsx
@@ -29,7 +29,7 @@ const meta: ComponentMeta< typeof NavigatorProvider > = {
 	},
 	parameters: {
 		controls: { expanded: true },
-		docs: { source: { state: 'open' } },
+		docs: { canvas: { sourceState: 'shown' } },
 	},
 };
 export default meta;