+++ packages/components/src/palette-edit/stories/index.story.tsx
@@ -20,7 +20,7 @@ const meta: ComponentMeta< typeof PaletteEdit > = {
 	parameters: {
 		actions: { argTypesRegex: '^on.*' },
 		controls: { expanded: true },
-		docs: { source: { state: 'open' } },
+		docs: { canvas: { sourceState: 'shown' } },
 	},
 };
 export default meta;