+++ packages/components/src/progress-bar/styles.ts
@@ -43,7 +43,7 @@ export const Indicator = styled.div< {
 	top: 0;
 	height: 100%;
 	border-radius: ${ CONFIG.radiusBlockUi };
-	background-color: var( --wp-components-color-accent, ${ COLORS.ui.theme } );
+	background-color: ${ COLORS.theme.accent };
 
 	${ ( { isIndeterminate, value } ) =>
 		isIndeterminate