+++ packages/components/src/radio-group/stories/index.story.js
@@ -39,7 +39,7 @@ export const _default = () => {
 	/* eslint-enable no-restricted-syntax */
 };
 
-export const disabled = () => {
+export const Disabled = () => {
 	/* eslint-disable no-restricted-syntax */
 	return (
 		<RadioGroup
@@ -77,6 +77,6 @@ const ControlledRadioGroupWithState = () => {
 	/* eslint-enable no-restricted-syntax */
 };
 
-export const controlled = () => {
+export const Controlled = () => {
 	return <ControlledRadioGroupWithState />;
 };