+++ packages/components/src/slot-fill/stories/index.story.js
@@ -13,7 +13,7 @@ export default {
 	component: Slot,
 	parameters: {
 		controls: { expanded: true },
-		docs: { source: { state: 'open' } },
+		docs: { canvas: { sourceState: 'shown' } },
 	},
 };
 
@@ -33,7 +33,7 @@ export const _default = () => {
 	);
 };
 
-export const withFillProps = () => {
+export const WithFillProps = () => {
 	return (
 		<SlotFillProvider>
 			<h2>Profile</h2>
@@ -61,7 +61,7 @@ export const withFillProps = () => {
 	);
 };
 
-export const withContext = () => {
+export const WithContext = () => {
 	const Context = createContext();
 	const ContextFill = ( { name } ) => {
 		const value = useContext( Context );