+++ packages/components/src/truncate/stories/index.story.tsx
@@ -19,7 +19,7 @@ const meta: ComponentMeta< typeof Truncate > = {
 		controls: {
 			expanded: true,
 		},
-		docs: { source: { state: 'open' } },
+		docs: { canvas: { sourceState: 'shown' } },
 	},
 };
 export default meta;