+++ packages/e2e-test-utils-playwright/src/admin/visit-site-editor.ts
@@ -66,5 +66,8 @@ export async function visitSiteEditor(
 	// TODO: Ideally the content underneath the spinner should be marked inert until it's ready.
 	await this.page
 		.locator( '.edit-site-canvas-spinner' )
-		.waitFor( { state: 'hidden' } );
+		// Bigger timeout is needed for larger entities, for example the large
+		// post html fixture that we load for performance tests, which often
+		// doesn't make it under the default 10 seconds.
+		.waitFor( { state: 'hidden', timeout: 60_000 } );
 }